requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-docx==1.1.2
hypothesis>=6.0.0
pytest>=7.0.0
